added basic generator for JSON data
Supports objects, nested objects, arrays, and all basic JSON types.

JSON strings can be also be generated using a regular expression.

Added a markov-chain sentence generator for random, believable contents
but it seems to expontially slow generation.

Also, it depends on chug, which it shouldn't. This was a hack to make
it work with forsight without having to futz with forsight much.